pyproject.toml
[tool.poetry]
name = "sayvai-graph-agents"
version = "0.0.1"
description = "A project"
authors = ["Sanjaypranav <sanjaypriya195@.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.43"
python-dotenv = "^1.0.1"
streamlit = {extras = ["all"], version = "^1.39.0"}
langchain-community = "^0.3.4"
langchain-openai = "^0.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
